actualizacion 29-06 detalle de inventario get y eliminar bodega que elimina todo

+++ api/post/bodega.php
@@ -10,6 +10,9 @@ include_once '../../config/conexion.php';
 include_once '../../Controller/controller_bodega.php';
 include_once '../../Controller/Controller_Trabajador.php';
 include_once '../../Controller/Controller_trabajador_has_bodega.php';
+include_once '../../Controller/Controller_detalle_inventario.php';
+include_once '../../Controller/Controller_Inventario.php';
+include_once '../../Controller/Controller_Gasto.php';
 
 $database = new conexion();
 $db = $database->connect();
@@ -41,9 +44,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     if ($validador == true) {
         if ($pos->Validar_tipo_trabajador($pos->id_tipo_trabajador) != null) { 
             if ($post->create()) {
-                echo json_encode(
-                    array('message' => 'Post Created')
-                );
+                
             if ($post->buscar_el_ultimo_id() != null) {
                 
                     //accion a trabajador as bodega (rut_trabajador,id_bodega)
@@ -52,6 +53,12 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                     if ($po->create_trabajador_has_bodega($rut_trabajador,$id_bodega)==false) {
                         echo json_encode(
                             array('message' => 'Error en ingreso de datos teniendo en cuenta el rut del trabajador y el codigo de la bodega')
+                        );
+                    }else {
+                        //crear gasto de esa bodega con margen 0 de gasto
+                        
+                        echo json_encode(
+                            array('message' => 'Post Created')
                         );
                     }
                 } else {
@@ -87,6 +94,9 @@ if ($_SERVER['REQUEST_METHOD'] == 'GET') {
                 array('message' => 'No existe datos sobre la bodega N°' . $post->id_bodega)
             );
         } else {
+            //obtenr los gastos de esa bodega y actualizar datos de estos 
+
+
             if ($post->read_single()) {
                 $post_item = array(
                     'id_bodega' => $post->id_bodega,
@@ -145,6 +155,10 @@ if ($_SERVER['REQUEST_METHOD'] == 'DELETE') {
     // Instiate blog post object
     $post = new controller_bodega($GLOBALS['db']);
     $po = new Controller_trabajador_has_bodega($GLOBALS['db']);
+    $i = new Controller_Inventario($GLOBALS['db']);
+    $di = new Controller_detalle_inventario($GLOBALS['db']);
+    $g = new Controller_Gasto($GLOBALS['db']);
+
     
     // GET ID
     $post->id_bodega = isset($_GET['id_bodega']) ? $_GET['id_bodega'] : die();
@@ -164,23 +178,41 @@ if ($_SERVER['REQUEST_METHOD'] == 'DELETE') {
             array('message' => 'no se puede eliminar esta bodega ya que esta relaciona a un gasto existente')
         );
     }
-
+//eliminar los productos de detalle inv , gastos y actulizar el total del inv.
     
     if ($validator == true) {
-        if ($po->delete_trabajador_has_bodega($post->id_bodega)) {
-            if ($post->delete_single()) {
-                echo json_encode(
-                    array('message' => 'Post deleted')
-                );
+        $valor_di = $di->buscardor_valor_producto_por_bodega($post->id_bodega);
+        //obtenr el valor actual del inventario
+        if ($i->Obtner_valor_inventario()==null) {
+            echo json_encode(
+                array('message' => 'Post not deleted')
+            );
+        }else {
+            $vi=$i->Obtner_valor_inventario();
+        }
+        $vi = $vi -$valor_di;
+        $i -> actualizar_valor($vi,1);
+        if ($g ->delete_single_gasto_por_bodega($post->id_bodega)==false) {
+            echo json_encode(
+                array('message' => 'No se elimino el gasto de la bodega')
+            );
+        }else {
+            if ($po->delete_trabajador_has_bodega($post->id_bodega)) {
+                if ($post->delete_single()) {
+                    
+                    echo json_encode(
+                        array('message' => 'Post deleted')
+                    );
+                } else {
+                    echo json_encode(
+                        array('message' => 'Post not deleted')
+                    );
+                }       
             } else {
                 echo json_encode(
                     array('message' => 'Post not deleted')
                 );
-            }       
-        } else {
-            echo json_encode(
-                array('message' => 'Post not deleted')
-            );
+            }
         }
     }
